+++ RedPandaIDE/mainwindow.h
@@ -794,9 +794,12 @@ private slots:
 
     void on_actionGNU_Assembler_Manual_triggered();
 
+#ifdef ARCH_X86_64
     void on_actionx86_Assembly_Language_Reference_Manual_triggered();
-
+#endif
+#ifdef ARCH_X86
     void on_actionIA_32_Assembly_Language_Reference_Manual_triggered();
+#endif
 
     void on_actionAdd_Watchpoint_triggered();